+++ sys/ufs/ffs/ffs_vfsops.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: ffs_vfsops.c,v 1.13 1995/03/08 01:51:49 cgd Exp $	*/
+/*	$NetBSD: ffs_vfsops.c,v 1.14 1995/03/09 12:06:08 mycroft Exp $	*/
 
 /*
  * Copyright (c) 1989, 1991, 1993, 1994
@@ -94,7 +94,7 @@ ffs_mountroot()
 	register struct mount *mp;
 	struct proc *p = curproc;	/* XXX */
 	struct ufsmount *ump;
-	u_long size;
+	size_t size;
 	int error;
 	
 	/*
@@ -149,7 +149,7 @@ ffs_mount(mp, path, data, ndp, p)
 	struct ufs_args args;
 	struct ufsmount *ump;
 	register struct fs *fs;
-	u_long size;
+	size_t size;
 	int error, flags;
 	mode_t accessmode;
